#16 Fix: index.js www.js 분리
