+++ bin/www.js
@@ -1,22 +1,6 @@
-const express = require('express');
-
-const app = express();
-const bodyParser = require('body-parser');
-const cookieParser = require('cookie-parser');
-const passport = require('passport');
-const router = require('../routes/index');
+const app = require('..');
+require('dotenv').config();
 
 const port = process.env.PORT || 5000;
 
-// parse application/x-www-form-urlencoded
-app.use(bodyParser.urlencoded({ extended: false }));
-
-// parse application/json
-app.use(bodyParser.json());
-app.use(cookieParser());
-app.use(passport.initialize());
-app.use(passport.session());
-
-app.use('/', router);
-
-app.listen(port, () => console.log(`Prod Server listening on port ${port}!`));
+app.listen(port, () => console.log(`Dev Server listening on port ${port}!`));
