+++ package.json
@@ -19,6 +19,7 @@
     "express-validator": "^6.11.1",
     "helmet": "^4.6.0",
     "jsonwebtoken": "^8.5.1",
+    "morgan": "^1.10.0",
     "mysql2": "^2.2.5",
     "node-fetch": "^2.6.1",
     "nodemailer": "^6.6.1",
